+++ src/main/java/pl/plajer/villagedefense/handlers/language/LanguageMigrator.java
@@ -42,7 +42,7 @@ public class LanguageMigrator {
   public static final int CONFIG_FILE_VERSION = 8;
   private Main plugin;
   private List<String> migratable = Arrays.asList(Constants.Files.CONFIG.getName(), Constants.Files.KITS.getName(),
-      Constants.Files.KITS.getName(), Constants.Files.LANGUAGE.getName(), Constants.Files.LOBBY_ITEMS.getName(), Constants.Files.MYSQL.getName());
+      Constants.Files.KITS.getName(), Constants.Files.LANGUAGE.getName(), Constants.Files.SPECIAL_ITEMS.getName(), Constants.Files.MYSQL.getName());
 
   public LanguageMigrator(Main plugin) {
     this.plugin = plugin;