+++ src/main/java/pl/plajer/villagedefense/handlers/language/Messages.java
@@ -104,7 +104,6 @@ public enum Messages {
   KITS_MENU_LOCKED_UNLOCK_IN_STORE("Kits.Kit-Menu.Locked-Lores.Unlock-In-Store"),
   KITS_NOT_UNLOCKED_MESSAGE("Kits.Not-Unlocked-Message"),
   KITS_CHOOSE_MESSAGE("Kits.Choose-Message"),
-  KITS_KIT_MENU_ITEM_NAME("Kits.Kit-Menu-Item-Name"),
   KITS_OPEN_KIT_MENU("Kits.Open-Kit-Menu"),
   KITS_ABILITY_STILL_ON_COOLDOWN("Kits.Ability-Still-On-Cooldown"),
 
@@ -244,10 +243,8 @@ public enum Messages {
   WOLF_SPAWNED("In-Game.Messages.Wolf-Spawned"),
   ZOMBIE_GOT_STUCK_IN_THE_MAP("In-Game.Messages.Zombie-Got-Stuck-In-The-Map"),
 
-  SPECTATOR_ITEM_NAME("In-Game.Spectator.Spectator-Item-Name"),
   SPECTATOR_MENU_NAME("In-Game.Spectator.Spectator-Menu-Name"),
   SPECTATOR_TARGET_PLAYER_HEALTH("In-Game.Spectator.Target-Player-Health"),
-  SPECTATOR_SETTINGS_MENU_ITEM_NAME("In-Game.Spectator.Settings-Menu.Item-Name"),
   SPECTATOR_SETTINGS_MENU_INVENTORY_NAME("In-Game.Spectator.Settings-Menu.Inventory-Name"),
   SPECTATOR_SETTINGS_MENU_SPEED_NAME("In-Game.Spectator.Settings-Menu.Speed-Name"),