+++ src/main/java/pl/plajer/villagedefense/kits/KitMenuHandler.java
@@ -22,22 +22,20 @@ import com.github.stefvanschie.inventoryframework.Gui;
 import com.github.stefvanschie.inventoryframework.GuiItem;
 import com.github.stefvanschie.inventoryframework.pane.StaticPane;
 
-import java.util.Arrays;
-
 import org.bukkit.Bukkit;
-import org.bukkit.Material;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.block.Action;
 import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.ItemMeta;
 
 import pl.plajer.villagedefense.Main;
 import pl.plajer.villagedefense.api.event.player.VillagePlayerChooseKitEvent;
 import pl.plajer.villagedefense.arena.Arena;
 import pl.plajer.villagedefense.arena.ArenaRegistry;
+import pl.plajer.villagedefense.handlers.items.SpecialItem;
+import pl.plajer.villagedefense.handlers.items.SpecialItemManager;
 import pl.plajer.villagedefense.handlers.language.Messages;
 import pl.plajer.villagedefense.kits.basekits.Kit;
 import pl.plajer.villagedefense.user.User;
@@ -46,50 +44,27 @@ import pl.plajerlair.commonsbox.minecraft.item.ItemBuilder;
 import pl.plajerlair.commonsbox.minecraft.item.ItemUtils;
 
 /**
- * Class for setting Kit values.
- * Need to use before registering new kit!
- *
- * @author TomTheDeveloper
+ * @author Plajer
+ * <p>
+ * Created at 12.07.2019
  */
-public class KitManager implements Listener {
+public class KitMenuHandler implements Listener {
 
-  private String[] description = new String[] {};
   private Main plugin;
-  private String itemName;
-  private Material material;
-  private String menuName;
-
   private String unlockedString;
   private String lockedString;
+  private SpecialItem kitItem;
 
-  public KitManager(Main plugin) {
+  public KitMenuHandler(Main plugin) {
     this.plugin = plugin;
-    itemName = plugin.getChatManager().colorMessage(Messages.KITS_KIT_MENU_ITEM_NAME);
+    this.kitItem = plugin.getSpecialItemManager().getSpecialItem(SpecialItemManager.SpecialItems.KIT_SELECTOR.getName());
     unlockedString = plugin.getChatManager().colorMessage(Messages.KITS_MENU_UNLOCKED_LORE);
     lockedString = plugin.getChatManager().colorMessage(Messages.KITS_MENU_LOCKED_LORE);
     plugin.getServer().getPluginManager().registerEvents(this, plugin);
   }
 
-  /**
-   * Returns name of kit
-   *
-   * @return name of kit
-   */
-  public String getItemName() {
-    return itemName;
-  }
-
-  /**
-   * Sets name of kit
-   *
-   * @param name kit name
-   */
-  public void setItemName(String name) {
-    this.itemName = name;
-  }
-
-  public void createKitMenu(Player player) {
-    Gui gui = new Gui(plugin, Utils.serializeInt(KitRegistry.getKits().size()) / 9, getMenuName());
+  public void createMenu(Player player) {
+    Gui gui = new Gui(plugin, Utils.serializeInt(KitRegistry.getKits().size()) / 9, plugin.getChatManager().colorMessage(Messages.KITS_OPEN_KIT_MENU));
     StaticPane pane = new StaticPane(9, gui.getRows());
     gui.addPane(pane);
     int x = 0;
@@ -113,6 +88,16 @@ public class KitManager implements Listener {
         }
         VillagePlayerChooseKitEvent event = new VillagePlayerChooseKitEvent(player, KitRegistry.getKit(e.getCurrentItem()), arena);
         Bukkit.getPluginManager().callEvent(event);
+        if (event.isCancelled()) {
+          return;
+        }
+        if (!kit.isUnlockedByPlayer(player)) {
+          player.sendMessage(plugin.getChatManager().colorMessage(Messages.KITS_NOT_UNLOCKED_MESSAGE).replace("%KIT%", kit.getName()));
+          return;
+        }
+        User user = plugin.getUserManager().getUser(player);
+        user.setKit(kit);
+        player.sendMessage(plugin.getChatManager().colorMessage(Messages.KITS_CHOOSE_MESSAGE).replace("%KIT%", kit.getName()));
       }), x, y);
       x++;
       if (x == 9) {
@@ -123,82 +108,17 @@ public class KitManager implements Listener {
     gui.show(player);
   }
 
-  /**
-   * Returns material represented by kit
-   *
-   * @return material represented by kit
-   */
-  public Material getMaterial() {
-    return material;
-  }
-
-  /**
-   * Sets material that kit will represents
-   *
-   * @param material material that kit will represents
-   */
-  public void setMaterial(Material material) {
-    this.material = material;
-  }
-
-  /**
-   * Returns description of kit
-   *
-   * @return description of kit
-   */
-  public String[] getDescription() {
-    return description.clone();
-  }
-
-  /**
-   * Sets description of kit
-   *
-   * @param description description of kit
-   */
-  public void setDescription(String[] description) {
-    this.description = description.clone();
-  }
-
-  public String getMenuName() {
-    return menuName;
-  }
-
-  public void setMenuName(String menuName) {
-    this.menuName = menuName;
-  }
-
-  public void giveKitMenuItem(Player player) {
-    ItemStack itemStack = new ItemStack(getMaterial());
-    ItemMeta itemMeta = itemStack.getItemMeta();
-    itemMeta.setDisplayName(getItemName());
-    itemMeta.setLore(Arrays.asList(getDescription()));
-    itemStack.setItemMeta(itemMeta);
-    player.getInventory().addItem(itemStack);
-  }
-
   @EventHandler
   public void onKitMenuItemClick(PlayerInteractEvent e) {
     if (!(e.getAction() == Action.RIGHT_CLICK_AIR || e.getAction() == Action.RIGHT_CLICK_BLOCK)) {
       return;
     }
     ItemStack stack = e.getPlayer().getInventory().getItemInMainHand();
-    if (stack.getType() != getMaterial() || !stack.hasItemMeta() || !stack.getItemMeta().hasLore()) {
+    if (!stack.equals(kitItem.getItemStack())) {
       return;
     }
-    if (!stack.getItemMeta().getDisplayName().equalsIgnoreCase(getItemName())) {
-      return;
-    }
-    createKitMenu(e.getPlayer());
+    e.setCancelled(true);
+    createMenu(e.getPlayer());
   }
 
-  @EventHandler
-  public void onKitChoose(VillagePlayerChooseKitEvent e) {
-    if (!e.getKit().isUnlockedByPlayer(e.getPlayer())) {
-      e.getPlayer().sendMessage(plugin.getChatManager().colorMessage(Messages.KITS_NOT_UNLOCKED_MESSAGE).replace("%KIT%", e.getKit().getName()));
-      return;
-    }
-    User user = plugin.getUserManager().getUser(e.getPlayer());
-    user.setKit(e.getKit());
-    e.getPlayer().sendMessage(plugin.getChatManager().colorMessage(Messages.KITS_CHOOSE_MESSAGE).replace("%KIT%", e.getKit().getName()));
-  }
 }