+++ src/main/java/pl/plajer/villagedefense/utils/constants/Constants.java
@@ -25,7 +25,7 @@ public class Constants {
 
   public enum Files {
     ARENAS("arenas"), CONFIG("config"), KITS("kits"), LANGUAGE("language"),
-    LOBBY_ITEMS("lobbyitems"), MYSQL("mysql"), REWARDS("rewards"), STATS("stats");
+    SPECIAL_ITEMS("special_items"), MYSQL("mysql"), REWARDS("rewards"), STATS("stats");
 
     private String name;